requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
channels==4.0.0
channels-redis==4.0.0
Django==4.1.6
django-rest-auth==0.9.5
djangorestframework==3.14.0
msgpack==1.0.4
python-decouple==3.7
pytz==2022.7.1
redis==4.4.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
